Practica comentada falta corregir error en la comprobacion de la recta

+++ src/8queens.js
@@ -11,7 +11,8 @@
  * Fecha: 8/03/2020
  *
  * Programa que muestra por pantalla las diferentes combinaciones en que se pueden
- * colocar 8 reinas en un tablero de ajedrez sin que se amenacen
+ * colocar 8 reinas en un tablero de ajedrez sin que se amenacen ademas comprobando
+ * si no pertenecen a la recta formada por dos reinas ya introducidas
  *
  * ---------------------------------------------------------------------------------------------------------------------*/
 
@@ -25,7 +26,7 @@ function constructor(size) {
   for (let i = 0; i < size; i++) {
     tablero.fila[i] = []
     for (let j = 0; j < size; j++) {
-        tablero.fila[i][j] = ".";
+      tablero.fila[i][j] = ".";
     }
 
     tablero.reinasEnFila[i] = false;
@@ -41,7 +42,7 @@ function imprimirTablero() {
     let cadena = "";
 
     for (let j = 0; j < tablero.fila[i].length; j++) {
-        cadena += " " + tablero.fila[i][j];
+      cadena += " " + tablero.fila[i][j];
     }
     console.log(cadena);
   }
@@ -56,11 +57,11 @@ function imprimirTablero() {
  */
 function insertarReina(columna) {
   if (tablero.reinasAlmacenadas.length === tablero.size) {
-      imprimirTablero();
+    imprimirTablero();
   }
   else {
     for (let fila = 0; fila < tablero.fila.length; fila++) {
-      if (comprobarFila(fila) && comprobarColumna(columna) && comprobarDiagonalNormal(fila, columna)) {
+      if (comprobarFila(fila) && comprobarColumna(columna) && comprobarDiagonalNormal(fila, columna) && comprobarLinea(fila, columna)) {
         crearReina(fila, columna);
         insertarReina(columna + 1);
         eliminarReina();
@@ -68,6 +69,7 @@ function insertarReina(columna) {
     }
   }
 }
+
 /**
  * @param  {} fila, coordenada en el eje X en la que se va a insertar la reina
  * @param  {} columna, coordenada en el eje Y en la que se va a insertar la reina
@@ -100,8 +102,58 @@ function eliminarReina() {
   tablero.reinasEnFila[ejeX] = false;
   tablero.reinasEnColumna[ejeY] = false;
   tablero.fila[ejeX][ejeY] = ".";
-
   tablero.reinasAlmacenadas.pop();
+    
+}
+
+/**
+ * @param  {} fila, coordenada en el eje X, de la reina que queremos introducir en la posicion dada
+ * @param  {} columna, coordenada en el eje Y, de la reina que queremos introducir en la posicion dada
+ * @description, Funcion utilizada para generar todas las combinaciones posibles de rectas que se pueden
+ * generar con las reinas ya introducidas en el tablero
+ */
+function comprobarLinea(fila, columna) {
+  if (tablero.reinasAlmacenadas.length < 2) {
+    return true;
+  }
+  else {
+    for (let reinaPrimera = 0; reinaPrimera < tablero.reinasAlmacenadas.length; reinaPrimera++) {
+      for (let reinaSegunda = reinaPrimera + 1; reinaSegunda < tablero.reinasAlmacenadas.length; reinaSegunda++) {
+        if (isPointInLine(fila, columna, lineFromTo(tablero.reinasAlmacenadas[reinaPrimera],
+          tablero.reinasAlmacenadas[reinaSegunda]))) {
+            return false;
+        }
+      }
+    }
+  }
+
+  return true;
+}
+
+/**
+ * @param  {} fila, coordena en el eje X, perteneciente a la reina que queremos colocar
+ * @param  {} columna, coordena en el eje Y, perteneciente a la reina que queremos colocar
+ * @param  {} linea, line formada por dos reinas dadas que utilizaremos para comprobar
+ * @description, Funcion usada para comprobar si un punto dado pertenece a la recta
+ * formada por dos reinas
+ */
+function isPointInLine(fila, columna, linea) {
+  return (((linea.pendiente * fila) + linea.desplazamiento) - columna) === 0;
+}
+
+/**
+ * @param  {} reinaPrimera, parametro uno, usado como primer punto que utilizaremos para
+ * trazar la recta entre los dos puntos
+ * @param  {} reinaSegunda, parametro dos, usado como segundo punto que utilizaremos para
+ * trazar la recta entre los dos puntos
+ * @description, Funcion utilizada para obtener los parametros que se utilizan para comprobar
+ * si un punto dado pertenece a la recta que forman dos reinas
+ */
+function lineFromTo(reinaPrimera, reinaSegunda) {
+  let pendiente_ = ((reinaSegunda.columna - reinaPrimera.columna) / (reinaSegunda.fila - reinaPrimera.fila));
+  let desplazamiento_ = reinaPrimera.columna + pendiente_ * reinaPrimera.fila;
+
+  return {pendiente: pendiente_, desplazamiento: desplazamiento_};
 }
 
 /**
@@ -158,8 +210,7 @@ function comprobarDiagonalNormal(fila, columna) {
     ejeY++;
   }
 
-  // Recorremos la diagonal del segundo cuadrante respecto a la posicion (ejeX negativo y
-  // ejeY positivo, 2º cuadrante)
+  // Recorremos la diagonal del segundo cuadrante respecto a la posicion (ejeX negativo y ejeY positivo, 2º cuadrante)
   ejeX = fila;
   ejeY = columna;
   while (ejeX >= 0 && ejeY >= 0) {
@@ -171,8 +222,7 @@ function comprobarDiagonalNormal(fila, columna) {
     ejeY--;
   }
 
-  // Recorremos la diagonal del tercer cuadrante respecto a la posicion (ejeX negativo y
-  // ejeY negativo, 3º cuadrante)
+  // Recorremos la diagonal del tercer cuadrante respecto a la posicion (ejeX negativo y ejeY negativo, 3º cuadrante)
   ejeX = fila;
   ejeY = columna;
   while (ejeX < tablero.fila.length && ejeY >= 0) {
@@ -184,8 +234,7 @@ function comprobarDiagonalNormal(fila, columna) {
     ejeY--;
   }
 
-  // Recorremos la diagonal del cuarto cuadrante respecto a la posicion (ejeX positivo y
-  // ejeY positivo, 4º cuadrante)
+  // Recorremos la diagonal del cuarto cuadrante respecto a la posicion (ejeX positivo y ejeY positivo, 4º cuadrante)
   ejeX = fila;
   ejeY = columna;
   while (ejeX < tablero.fila.length && ejeY < tablero.fila.length) {
